Fixed "Auto Unlock"
The functionality in this script to set the "auto unlock" feature of Bitlocker for drives other than the system drive (C usually) wasn't implemented correctly. Mostly just a dumb variable mistake where I used the wrong variable so it was trying to set auto unlock on a NULL value. Fixed now and tested/working!