Updated to consider reg_sz character limitations
I was updating per volumeid in reg_sz entries but noticed that if the same volume is encrypted/decrypted multiple times I was keeping each version of that... so it could fill the character limit of that. There's really no reason to keep a previous key of the exact same volumeid, so just overwriting to have the single latest key for the unique volumeid is more than sufficient.

Additionally I added an `Any` option for the type of encryption since I've found some people don't care about the encryption method as long as it's encrypted at all.